+++ apps/web/src/app/[locale]/(optional-space)/new/loading.tsx
