+++ apps/web/src/app/[locale]/(optional-space)/new/page.tsx
