+++ apps/web/src/app/[locale]/(optional-space)/poll/[urlId]/guest-poll-alert.tsx
