+++ apps/web/src/app/[locale]/(optional-space)/poll/[urlId]/layout.tsx
